Data table read logic
Created dev level to work through logic required to access data table. SelLevelAttributes is used to grab and update information form the GameInstance and then send this to the speakers. Word logic is the current dev BP to work out how to use this information from the data tables but will eventually become a speaker