Icon appear bug fixes
Now check for visibility at start and base fade-in/out animations on current visibility state to make sure fad in/out does not switch (ie disappears when looking at the icon direction instead of appearing).
Now calculates the animation time for the fad-in/out so that this time can be used to start a fad-in/out. For example when quickly transitioning between the fadin + fade out, you want to start the fade in at the right time of the fade out or you will get jumps in transparency.